refactor car
